+++ feature/News/src/LayerNews/AndroidManifest.xml
@@ -5,10 +5,16 @@
 
     <application>
         <activity
-            android:name=".presentation.view.DetailsNewsActivity"
+            android:name=".presentation.view.FavoriteActivity"
+            android:exported="false" />
+        <activity
+            android:name=".presentation.view.FiltersActivity"
             android:exported="false" />
         <activity
             android:name=".presentation.view.HomeActivity"
+            android:exported="false" />
+        <activity
+            android:name=".presentation.view.DetailsNewsActivity"
             android:exported="false" />
         <provider
             android:name="androidx.core.content.FileProvider"
