--- feature/News/src/LayerNews/java/com/example/feature/news/data/remote/NewsResponse.kt
@@ -1,7 +1,0 @@
-package com.example.feature.news.data.remote
-
-data class NewsResponse(
-    val status: String,
-    val totalResults: Int,
-    val articles: List<NewsDto>
-)
