--- feature/News/src/LayerNews/java/com/example/feature/news/di/NewsModule.kt
@@ -1,21 +1,0 @@
-package com.example.feature.news.di
-
-import com.example.feature.news.data.repository.NewsRepositoryImpl
-import com.example.feature.news.domain.repository.NewsRepository
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-abstract class NewsModule {
-
-    @Binds
-    @Singleton
-    abstract fun bindNewsRepository(
-        impl: NewsRepositoryImpl
-    ): NewsRepository
-}
-
