--- feature/News/src/LayerNews/java/com/example/feature/news/di/Qualifiers.kt
@@ -1,7 +1,0 @@
-package com.example.feature.news.di
-
-import javax.inject.Qualifier
-
-@Qualifier
-@Retention(AnnotationRetention.BINARY)
-annotation class NewsRetrofit
