--- feature/News/src/LayerNews/java/com/example/feature/news/domain/model/News.kt
@@ -1,9 +1,0 @@
-package com.example.feature.news.domain.model
-
-data class News(
-    val id: Int,
-    val title: String,
-    val description: String,
-    val imageUrl: String,
-    val date: String
-)
