--- feature/News/src/LayerNews/java/com/example/feature/news/domain/usecase/GetTopHeadlinesUseCase.kt
@@ -1,13 +1,0 @@
-package com.example.feature.news.domain.usecase
-
-import com.example.feature.news.domain.model.NewsResponse
-import com.example.feature.news.domain.repository.NewsRepository
-import javax.inject.Inject
-
-class GetTopHeadlinesUseCase @Inject constructor(
-    private val newsRepository: NewsRepository
-) {
-    suspend operator fun invoke(country: String = "us", page: Int = 1): Result<NewsResponse> {
-        return newsRepository.getTopHeadlines(country, page)
-    }
-}
