+++ feature/News/src/LayerNews/java/com/example/feature/news/presentation/view/favorites/FavoritesActivity.kt
