+++ feature/News/src/LayerNews/java/com/example/feature/news/presentation/viewModel/DetailNewsViewModel.kt
@@ -1,49 +1,54 @@
 package com.example.feature.news.presentation.viewModel
 
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
 import com.example.feature.news.domain.model.Article
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.firestore.FirebaseFirestore
+import com.example.feature.news.domain.usecase.CheckIsFavoriteUseCase
+import com.example.feature.news.domain.usecase.ToggleFavoriteUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
 class DetailNewsViewModel @Inject constructor(
-    private val firestore: FirebaseFirestore,
-    private val auth: FirebaseAuth
+    private val toggleFavoriteUseCase: ToggleFavoriteUseCase,
+    private val checkIsFavoriteUseCase: CheckIsFavoriteUseCase
 ) : ViewModel() {
+
     private val _isFavorite = MutableStateFlow(false)
     val isFavorite = _isFavorite.asStateFlow()
 
+    private val _error = MutableStateFlow<String?>(null)
+    val error = _error.asStateFlow()
+
     fun toggleFavorite(article: Article) {
-        val userId = auth.currentUser?.uid ?: return
-        val newState = !_isFavorite.value
-        _isFavorite.value = newState
-
-        val documentId = article.url.hashCode().toString()
-        val docRef = firestore.collection("users").document(userId)
-            .collection("favorites").document(documentId)
-
-        if (newState) {
-            docRef.set(article)
-                .addOnFailureListener { _isFavorite.value = false }
-        } else {
-            docRef.delete()
-                .addOnFailureListener { _isFavorite.value = true }
+        viewModelScope.launch {
+            toggleFavoriteUseCase(article).fold(
+                onSuccess = { newState ->
+                    _isFavorite.value = newState
+                    _error.value = null
+                },
+                onFailure = { exception ->
+                    _error.value = exception.message ?: "Erro ao atualizar favorito"
+                }
+            )
         }
     }
 
     fun checkIfIsFavorite(articleUrl: String) {
-        val userId = auth.currentUser?.uid ?: return
-        val documentId = articleUrl.hashCode().toString()
-
-        firestore.collection("users").document(userId)
-            .collection("favorites").document(documentId)
-            .get()
-            .addOnSuccessListener { document ->
-                _isFavorite.value = document.exists()
-            }
+        viewModelScope.launch {
+            checkIsFavoriteUseCase(articleUrl).fold(
+                onSuccess = { isFav ->
+                    _isFavorite.value = isFav
+                    _error.value = null
+                },
+                onFailure = { exception ->
+                    _error.value = exception.message ?: "Erro ao verificar favorito"
+                    _isFavorite.value = false
+                }
+            )
+        }
     }
 }
