+++ feature/News/src/LayerNews/res/layout/activity_favorites.xml
