+++ feature/News/src/LayerNews/res/layout/activity_filters.xml
