+++ feature/News/src/main/java/com/example/feature/news/domain/model/NewsModel.kt
@@ -18,8 +18,8 @@ data class Article(
 
 @Parcelize
 data class Source(
-    val id: String?,
-    val name: String
+    val id: String? = null,
+    val name: String = ""
 ): Parcelable
 
 data class NewsResponse(
